+++ src/blocks/accordion-group/attributes.ts
@@ -13,6 +13,11 @@ export const attributes = {
 		type: "string",
 		default: "",
 	},
+	headerElement: {
+		type: "string",
+		enum: ["h2", "h3", "h4", "h5", "h6", "p"],
+		default: "h2",
+	},
 } as const satisfies BlockAttributes;
 export type Attributes = typeof attributes;
 export type InterpretedAttributes = InterpretAttributes<Supports, Attributes>;