+++ src/blocks/accordion-group/block.json.ts
@@ -2,6 +2,7 @@ import type { Attributes } from "./attributes";
 import type { Supports } from "./supports";
 import type { BlockMetaData } from "@atomicsmash/blocks-helpers";
 import { attributes } from "./attributes";
+import { providesContext, usesContext } from "./context";
 import { example } from "./example";
 import { supports } from "./supports";
 
@@ -20,5 +21,8 @@ export const blockJson = {
 	attributes,
 	example,
 	supports,
+	providesContext,
+	usesContext,
+	allowedBlocks: ["launchpad-blocks/accordion"],
 } satisfies BlockMetaData<Supports, Attributes>;
 export default blockJson;