+++ src/blocks/accordion-group/edit.tsx
@@ -3,16 +3,72 @@ import type { CreateBlockEditProps } from "@atomicsmash/blocks-helpers";
 import {
 	useBlockProps,
 	InspectorControls,
-	InnerBlocks,
+	BlockControls,
+	Inserter,
+	store as blockEditorStore,
+	useInnerBlocksProps,
 } from "@wordpress/block-editor";
-import { CheckboxControl, Panel, PanelBody } from "@wordpress/components";
-import { useUniqueBlockId } from "../helpers.editor";
+import {
+	CheckboxControl,
+	Panel,
+	PanelBody,
+	ButtonGroup,
+	Button,
+	ToolbarGroup,
+} from "@wordpress/components";
+import { useSelect } from "@wordpress/data";
+import { blockDefault } from "@wordpress/icons";
+import { useUniqueBlockId, HeadingLevelSelect } from "../helpers.editor";
+import { attributes as definedAttributeOptions } from "./attributes";
 
 export type BlockEditProps = CreateBlockEditProps<InterpretedAttributes>;
 
-export function Edit({ clientId, attributes, setAttributes }: BlockEditProps) {
-	const { isMultiple } = attributes;
+export function Edit({
+	clientId,
+	attributes,
+	setAttributes,
+	isSelected,
+}: BlockEditProps) {
+	const { isMultiple, headerElement } = attributes;
 	const blockProps = useBlockProps();
+	const innerBlockProps = useInnerBlocksProps(blockProps, {
+		template: [
+			[
+				"launchpad-blocks/accordion",
+				{},
+				[
+					[
+						"core/paragraph",
+						{
+							placeholder:
+								"Add accordion content here. Type / to choose a block.",
+						},
+					],
+				],
+			],
+		],
+		renderAppender: () => (
+			<Inserter
+				rootClientId={clientId}
+				renderToggle={({ onToggle }: { onToggle: () => void }) => {
+					if (!isSelected && !isInnerBlockSelected) {
+						return null;
+					}
+					return (
+						<ButtonGroup>
+							<Button
+								className="accordion-inserter-button is-primary"
+								onClick={onToggle}
+							>
+								{blockDefault} Add new accordion to this group
+							</Button>
+						</ButtonGroup>
+					);
+				}}
+				isAppender
+			/>
+		),
+	});
 
 	useUniqueBlockId(
 		attributes,
@@ -20,6 +76,16 @@ export function Edit({ clientId, attributes, setAttributes }: BlockEditProps) {
 		clientId,
 		setAttributes,
 		"launchpad-blocks/accordion-group",
+	);
+
+	const isInnerBlockSelected = useSelect(
+		(select) =>
+			(
+				select(blockEditorStore) as {
+					hasSelectedInnerBlock: (clientId: string, deep: boolean) => boolean;
+				}
+			).hasSelectedInnerBlock(clientId, true),
+		[clientId],
 	);
 
 	return (
@@ -39,9 +105,18 @@ export function Edit({ clientId, attributes, setAttributes }: BlockEditProps) {
 					</PanelBody>
 				</Panel>
 			</InspectorControls>
-			<div {...blockProps}>
-				<InnerBlocks allowedBlocks={["launchpad-blocks/accordion"]} />
-			</div>
+			<BlockControls>
+				<ToolbarGroup>
+					<HeadingLevelSelect
+						levelOptions={definedAttributeOptions.headerElement.enum}
+						selectedLevel={headerElement}
+						setSelectedHeadingLevel={(newHeadingLevel) => {
+							setAttributes({ headerElement: newHeadingLevel });
+						}}
+					/>
+				</ToolbarGroup>
+			</BlockControls>
+			<div {...innerBlockProps} />
 		</>
 	);
 }