+++ src/blocks/accordion-group/tests/accordion.setup.ts
@@ -2,7 +2,11 @@ import { existsSync } from "node:fs";
 import { WordPressAdminInteraction } from "@atomicsmash/wordpress-tests-helper";
 import { test as setup } from "@playwright/test";
 import { CURRENT_WORDPRESS_VERSION } from "@tests/playwright-utils";
-import { accordionGroupV1, accordionGroupV2 } from "./fixture";
+import {
+	accordionGroupV1,
+	accordionGroupV2,
+	accordionGroupV3,
+} from "./fixture";
 import { contentPersistLocation, doTearDown } from "./index";
 
 setup("Accordion test setup", async ({ page, baseURL }) => {
@@ -27,6 +31,11 @@ setup("Accordion test setup", async ({ page, baseURL }) => {
 			title: "Accordion v2 tests",
 			postType: "page",
 		},
+		{
+			postIdentifier: "accordion-v3-test-page",
+			title: "Accordion v3 tests",
+			postType: "page",
+		},
 	]);
 	await adminHelper.addContentByFixture("accordion-v1-test-page", [
 		accordionGroupV1.allowMultiple(baseURL ?? ""),
@@ -35,5 +44,9 @@ setup("Accordion test setup", async ({ page, baseURL }) => {
 	await adminHelper.addContentByFixture("accordion-v2-test-page", [
 		accordionGroupV2.allowMultiple,
 		accordionGroupV2.doesNotAllowMultiple,
+	]);
+	await adminHelper.addContentByFixture("accordion-v3-test-page", [
+		accordionGroupV3.allowMultiple,
+		accordionGroupV3.doesNotAllowMultiple,
 	]);
 });