+++ src/blocks/accordion/block.json.ts
@@ -2,6 +2,7 @@ import type { Attributes } from "./attributes";
 import type { Supports } from "./supports";
 import type { BlockMetaData } from "@atomicsmash/blocks-helpers";
 import { attributes } from "./attributes";
+import { providesContext, usesContext } from "./context";
 import { supports } from "./supports";
 
 export const blockJson = {
@@ -19,5 +20,7 @@ export const blockJson = {
 	render: "file:./render.php",
 	attributes,
 	supports,
+	providesContext,
+	usesContext,
 } satisfies BlockMetaData<Supports, Attributes>;
 export default blockJson;