+++ src/blocks/accordion/index.tsx
@@ -1,4 +1,5 @@
 import type { InterpretedAttributes } from "./attributes";
+import type { InterpretedUsedContext } from "./context";
 import { registerBlockType } from "@atomicsmash/blocks-helpers";
 import { registerLaunchpadBlocksCollection } from "@launchpadBlocks/helpers.editor";
 import { Icon } from "@launchpadBlocks/svgs";
@@ -7,10 +8,15 @@ import { deprecated } from "./deprecation";
 import { Edit } from "./edit";
 import { save } from "./save";
 
-registerBlockType<InterpretedAttributes>(blockMetaData.name, {
-	icon: <Icon iconName="accordion-single" />,
-	edit: Edit,
-	save: save({ hasInnerBlocks: true }),
-	deprecated,
-});
+export { type InterpretedProvidesContext } from "./context";
+
+registerBlockType<InterpretedAttributes, InterpretedUsedContext>(
+	blockMetaData.name,
+	{
+		icon: <Icon iconName="accordion-single" />,
+		edit: Edit,
+		save: save({ hasInnerBlocks: true }),
+		deprecated,
+	},
+);
 registerLaunchpadBlocksCollection();